+++ yaqd-core/yaqd_core/_has_transformed_position.py
@@ -30,8 +30,8 @@ class HasTransformedPosition(HasLimits, HasPosition, IsDaemon):
         (i.e. in relative coordinates reference position is zero).
 
         This placeholder function has trivial behavior, f(x) = x.
-        Daemons that need more than a reference_position should overload this method.
-        Note that overloads should still obey inversion:
+        Daemons that need more than a reference_position should override this method.
+        Note that override should still obey inversion:
         ```
         _relative_to_transformed(_transformed_to_relative(x)) == x
         ```.
@@ -44,8 +44,8 @@ class HasTransformedPosition(HasLimits, HasPosition, IsDaemon):
         (i.e. in relative coordinates reference position is zero).
 
         This placeholder function has trivial behavior, f(x) = x.
-        Daemons that need more than a reference_position should overload this method.
-        Note that overloads should still obey inversion:
+        Daemons that need more than a reference_position should override this method.
+        Note that override should still obey inversion:
         ```
         _relative_to_transformed(_transformed_to_relative(x)) == x
         ```.